+++ hata/discord/auto_moderation/__init__.py
@@ -1,5 +1,11 @@
+from .action import *
 from .preinstanced import *
+from .rule import *
+from .trigger_metadata import *
 
 __all__ = (
+    *action.__all__,
     *preinstanced.__all__,
+    *rule.__all__,
+    *trigger_metadata.__all__,
 )